requirements.txt
google-generativeai>=0.8
sqlglot>=25
pandas>=2.2
numpy>=1.26
fastapi>=0.115
uvicorn>=0.30
sqlalchemy>=2.0
psycopg2-binary>=2.9
pymysql>=1.1
pyodbc>=5.0
aiosqlite>=0.20
sse-starlette>=2.1
slowapi>=0.1.9
python-dotenv
pyyaml>=6.0
httpx>=0.27
pytest-asyncio>=0.23
python-dateutil
python-multipart
gunicorn
qdrant_client